Auth fixes, cors config

+++ src/api/auth/services/logout.service.ts
@@ -1,5 +1,9 @@
 import { SessionModel } from '@/db/models/session.js';
 
-export async function logoutService(sessionId: string) {
+export async function logoutService(sessionId: string | undefined) {
+  if (!sessionId) {
+    return;
+  }
+
   await SessionModel.deleteOne({ _id: sessionId });
 }
